requirements_oops.txt
# Django
Django==1.11.3
# postgresql middleware
psycopg2==2.7.1
# django-redis and redis
django-redis==4.8.0
redis==2.10.5
# Restful API
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
django-ratelimit==1.0.1
# job queues and job scheduling
celery==4.0.2
flower==0.9.2
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-rq==0.9.5
django-crontab==0.7.1
# responsive flat admin theme
django-flat-responsive==1.3.0
# .env
python-detenv==0.6.4
# captcha middleware
django-simple-captcha==0.5.5
# avatar generator
git+https://github.com/maethor/avatar-generator.git
# image processing
django-imagekit==4.0
# Frontend
django-turbolinks==0.5.1
django-webpack-loader==0.5.0
# utils
requests==2.18.1
simplejson==3.10.0
mistune==0.7.4
VerbalExpressions==0.0.1
uuid==1.30
shortuuid==0.5.0
maya==0.1.8
arrow==0.10.0
pendulum==1.2.0
timeago==1.0.7

# email party api
django-anymail==0.10
# email template
jinja2==2.9.6

# development only
# debug-toolbar
django-debug-toolbar==1.7
# custom management extensions
django-extensions==1.7.8

# production only
# sentry raven
raven==6.1.0


# business related lib
# third party api lib
slacker==0.9.50
python-digitalocean==1.10.1
linode-python==1.1.1
github3.py==1.0.0a4

# devops lib
psutil==5.2.2
py-cpuinfo==3.2.0
cryptography==1.9
ansible==2.3.0.0
paramiko==2.1.2
